second commit of nov 22

+++ springSecure/src/main/java/com/springSecure/springSecure/repository/UserRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepo extends JpaRepository<UserObj,Long> {
+public interface UserRepo extends JpaRepository<UserObj,Integer> {
    UserObj findByName(String name);
     UserObj findByEmail(String email);
     boolean existsByName(String name);
